requirements.txt
amqp==5.1.1
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.2
billiard==4.1.0
celery==5.3.1
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.5
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.0
Django==4.1.10
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-timezone-field==5.1
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.4.0
exceptiongroup==1.1.2
Faker==19.1.0
h11==0.14.0
httpcore==0.17.3
idna==3.4
kombu==5.3.1
prompt-toolkit==3.0.39
pymongo==4.4.1
python-crontab==3.0.0
python-dateutil==2.8.2
pytz==2023.3
redis==4.6.0
requests==2.31.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.2.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
vine==5.0.0
wcwidth==0.2.6
